backend/requirements.txt
Django==2.2.7
django-tinymce4-lite==1.7.5
jsmin==2.2.2
pytz==2019.3
sqlparse==0.3.0
djangorestframework==3.14.0
django-cors-headers==4.1.0
